+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - added server config option to show locked handcrafted recipes
 - added server config option to isolate crafting tabs (remove the buttons to switch to another crafting tab)
+- add "is_consumed" boolean to (itemStack)ingredients ("true" by default). If set to "false", the ingredient is not consumed upon crafting.
+
+## Changes
+
+- reworked normal ingredients to "rpg_ingredients"
+- renamed "itemStackIngredients" into "rpg_item_stack_ingredients"
 
 # 1.3.0